Clean up O[U]StringBuffer ctors taking a length parameter
For !LIBO_INTERNAL_ONLY, go back to the state before
d91d98a53612a972de368186415aa48698e074d9 "fix gcc-4.0.1 and/or 32bit build for
string literals" of having just a sal_Int32 overload.  (For !LIBO_INTERNAL_ONLY,
the current state was confusing, as the existence of further overloads depended
on __cplusplus >= 201103L.)

And for LIBO_INTERNAL_ONLY:

* Have overloads for all integral types, with an assert checking that they do
  not use out-of-bounds length arguments.  (The std::make_unsigned_t dance in
  the assert is needed to avoid signed/unsigned mismatch warnings as seen at
  least with MSVC.)  This removes the need for explicitly casting arguments that
  are larger than (unsigned) int.  (And I cleaned up two such places that I had
  become aware of with the abandoned previous attempt
  <https://gerrit.libreoffice.org/c/core/+/132825> "Let O[U]StringBuffer ctors
  take sal_Int32 length parameters"; there may be more places to clean up,
  though.)

* Have deleted overloads for all integral types that are not actually integer
  types.  (This excludes signed char and unsigned char, as at least

>                             OUStringBuffer aBuf( rToken.GetByte());

  in ScTokenConversion::ConvertToTokenSequence in
  sc/source/ui/unoobj/tokenuno.cxx does a legitimate call with an argument of
  type sal_Int8 aka signed char.)

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132951
Tested-by: Jenkins

+++ desktop/source/app/cmdlineargs.cxx
@@ -219,7 +219,7 @@ CommandLineEvent CheckWebQuery(/* in,out */ OUString& arg, CommandLineEvent curE
         if (!SkipNewline(pPos))
             return curEvt;
 
-        OStringBuffer aResult(static_cast<unsigned int>(nRead));
+        OStringBuffer aResult(nRead);
         do
         {
             const char* pPos1 = pPos;
